requirements.txt
# requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
lxml==6.0.0
selenium==4.34.0
fake-useragent==2.2.0
python-dateutil==2.9.0.post0